requirements.txt
discord.py>=2.7.1,<3.0
PyNaCl>=1.5.0
davey
httpx>=0.27.0
python-dotenv>=1.2.2
